kernel: Force as many kernel tasks as possible onto CPU0
CPU0 is the boot CPU and more or less serves as the housekeeping CPU for
the system. Since CPU0 will always be busy with housekeeping (such as
servicing nearly all of the interrupts), we may as well tack on the rest
of the kernel's workload to CPU0 to gain some power savings (since we can
allow other CPUs to spend more time idling while CPU0 takes care of
frequent, low-overhead kernel work).

This makes workers that aren't scheduled to any particular CPU get
scheduled onto CPU0. Furthermore, all of the kthreads that haven't
been explicitly bound to a certain CPU are forced onto CPU0 as well.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

Conflicts:
	kernel/workqueue.c